Pedidos-carrito

+++ SGM/SGM.py
@@ -12,12 +12,15 @@ from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship, sessionmaker,declarative_base
 import os
 import json
+from sqlalchemy.dialects.mysql import INTEGER
+from datetime import datetime, time
+
 
 app = Flask(__name__)
 
 
 UPLOAD_FOLDER = 'SGM/static/uploads'
-ALLOWED_EXTENSIONS = {'png', 'jpg', 'jpeg', 'gif'}
+ALLOWED_EXTENSIONS = {'png', 'jpg', 'jpeg', 'gif','jfif','webp'}
 
 if not os.path.exists(UPLOAD_FOLDER):
     os.makedirs(UPLOAD_FOLDER)
@@ -73,7 +76,8 @@ class Element(db.Model):
     IDelement = Column(Integer, primary_key=True, autoincrement=True)
     name = Column(String(255), nullable=False)
     disp = db.Column(db.Boolean, nullable=True)
-    cant = Column(Integer, nullable=False)
+    cantotal = Column(Integer, nullable=False)
+    candisp = Column(INTEGER(unsigned=True), nullable=False)
     adds = Column(String(255), nullable=True)
     img = Column(String(255), nullable=True)
 
@@ -97,13 +101,13 @@ class Maintenance(db.Model):
     IDmaintenance = Column(Integer, primary_key=True, autoincrement=True)
     date = Column(Date, nullable=False)
     details = Column(String(255), nullable=True)
-    repeat = Column(Integer, nullable=False)
+    repeat = db.Column(db.Boolean, nullable=True)
     state = Column(String(255), nullable=False)
 
 class MaintenanceElement(db.Model):
     __tablename__ = 'maintenance_elm'
-   
-    IDmantenance = Column(Integer, ForeignKey('maintenance.IDmaintenance'), primary_key=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    IDmantenance = Column(Integer, ForeignKey('maintenance.IDmaintenance'))
     IDelement = Column(Integer, ForeignKey('elements.IDelement'))
     revised = Column(Integer, nullable=True)
     details = Column(String(255), nullable=True)
@@ -112,24 +116,28 @@ class Order(db.Model):
     __tablename__ = 'orders'
    
     IDorder = Column(Integer, primary_key=True, autoincrement=True)
+    IDuser = Column(Integer, ForeignKey('user.IDuser'))
     time = Column(Time, nullable=False)
     date = Column(Date, nullable=False)
     location = Column(String(255), nullable=False)
-    repeat = Column(Integer, nullable=False)
+    repeat = db.Column(db.Boolean, nullable=True)
     finish = Column(Date, nullable=True)
     state = Column(String(255), nullable=False)
 
 class OrderElement(db.Model):
     __tablename__ = 'orders_elm'
-   
-    IDorder = Column(Integer, ForeignKey('orders.IDorder'), primary_key=True)
-    IDelement = Column(Integer, ForeignKey('elements.IDelement'), primary_key=True)
+
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    IDorder = Column(Integer, ForeignKey('orders.IDorder'))
+    IDelement = Column(Integer, ForeignKey('elements.IDelement'))
+    cant = Column(Integer, nullable=False)
 
 class OrderRep(db.Model):
     __tablename__ = 'orders_rep'
-   
-    IDorder = Column(Integer, ForeignKey('orders.IDorder'), primary_key=True)
-    repeat_day = Column(String(255), primary_key=True)
+
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    IDorder = Column(Integer, ForeignKey('orders.IDorder'))
+    repeat_day = Column(String(255))
 
 class Location(db.Model):
     __tablename__ = 'location'
@@ -172,13 +180,13 @@ class RegisterForm(FlaskForm):
                            InputRequired(), Length(min=4, max=20)], render_kw={"placeholder": "Usuario"})
     
     email = StringField(validators=[
-                            InputRequired(), Length(min=4, max=45)], render_kw={"placeholder": "Email"})
+                            Length(max=45)], render_kw={"placeholder": "Email"})
 
     password = PasswordField(validators=[
                             InputRequired(), Length(min=8, max=20)], render_kw={"placeholder": "Contraseña"})
     
     
-    rol = SelectField('Roles', choices=[('ADMIN'), ('LV1'), ('LV2'), ('LVL3')])
+    rol = SelectField('Roles', choices=[('ADMIN'), ('LVL1'), ('LVL2'), ('LVL3'), ('LVL4')])
 
     submit = SubmitField('Registrar')
 
@@ -206,7 +214,7 @@ def validate_username(username):
 
 
 ############################ Login ###################################################
-@app.route('/')
+@app.route('/SGM')
 def home():
     return redirect(url_for('login'))
 
@@ -215,15 +223,25 @@ def home():
 def inicio():
     return render_template('inicio.html')
 
+@app.route("/pedidos", methods=['GET', 'POST'])
+@login_required
+def pedidos():
+    return render_template('PEDIDOS/pedidos.html')
+
 @app.route('/login', methods=['GET', 'POST'])
 def login():
     form = LoginForm()
     if form.validate_on_submit():
         user = User.query.filter_by(username=form.username.data).first()
         if user:
-            if bcrypt.check_password_hash(user.userpass, form.password.data):
-                login_user(user)
-                return redirect(url_for('inicio'))
+            if user.userclass != 'LVL1':
+                if bcrypt.check_password_hash(user.userpass, form.password.data):
+                    login_user(user)
+                    return redirect(url_for('inicio'))
+            if user.userclass == 'LVL1' or 'LVL2':
+                if bcrypt.check_password_hash(user.userpass, form.password.data):
+                    login_user(user)
+                    return redirect(url_for('pedidos'))
     return render_template('INICIO-SESION/index.html', form=form)
 
 
@@ -252,27 +270,28 @@ def logout():
 ############################ funciones de ADMIN ###################################################
 
 ########### registro #########
-@ app.route('/register', methods=['GET', 'POST'])
-def register():
+@ app.route('/registro', methods=['GET', 'POST'])
+def registro():
     form = RegisterForm()
     users = User.query.all()
     if form.validate_on_submit():
         if validate_username(form.username):
             flash('El usuario ya existe.', 'warning')
-            return redirect(url_for('register'))  
+            return redirect(url_for('registro'))  
         hashed_password = bcrypt.generate_password_hash(form.password.data)
         new_user = User(username=form.username.data, userpass=hashed_password,userclass=form.rol.data)
         db.session.add(new_user)
+        if form.email.data:
+            userid = User.query.filter(
+                User.username == form.username.data
+            ).first()
 
-        userid = User.query.filter(
-            User.username == form.username.data
-        ).first()
-
-        new_mail= Email(IDuser= userid.IDuser,address=form.email.data)
-        db.session.add(new_mail)
+            new_mail= Email(IDuser= userid.IDuser,address=form.email.data)
+            db.session.add(new_mail)
+            pass
         db.session.commit()
         flash('Usuario creado.', 'success')
-        return redirect(url_for('register'))
+        return redirect(url_for('registro'))
     return render_template('REGISTRO/index.html', users=users,form=form)
 
 
@@ -280,10 +299,14 @@ def register():
 @app.route('/borrar_usuario/<int:user_id>', methods=['POST'])
 def borrar_usuario(user_id):
     user = User.query.get_or_404(user_id)
+    
     db.session.delete(user)
+    mail = Email.query.get(user_id)
+    if mail:
+        db.session.delete(mail)
     db.session.commit()
     flash('Usuario eliminado exitosamente!', 'success')
-    return redirect(url_for('register'))
+    return redirect(url_for('registro'))
 
 ########### /////////// #########
 
@@ -332,7 +355,7 @@ def crear_categoria():
 def borrar_category(IDcategory):
 
     try:
-        elm_cat = ElementCategory.query.get_or_404(IDcategory)
+        elm_cat = ElementCategory.query.get(IDcategory)
         db.session.delete(elm_cat)
     except:
         pass
@@ -369,7 +392,7 @@ def subir_elemento():
     image.save(image_path)
     image_path='static/uploads/'+filename
     # Crear el nuevo elemento
-    nuevo_elemento = Element(name=name,disp=True, cant=quantity, adds=description, img=filename)
+    nuevo_elemento = Element(name=name,disp=True, cantotal=quantity,candisp=quantity, adds=description, img=filename)
 
     # Añadir a la base de datos
     db.session.add(nuevo_elemento)
@@ -390,16 +413,21 @@ def subir_elemento():
 
 @app.route('/borrar_elemento/<int:IDelement>', methods=['POST'])
 def borrar_elemento(IDelement):
+ 
     element = Element.query.get_or_404(IDelement)
-    elm_car = ElementCategory.query.get_or_404(IDelement)
+
+    elm_car = ElementCategory.query.get(IDelement)
+
     if element.img:
                     try:
                         os.remove(os.path.join(app.config['UPLOAD_FOLDER'], element.img))
                     except FileNotFoundError:
                         pass 
-    
-    db.session.delete(element)
-    db.session.delete(elm_car)
+    if element:
+        db.session.delete(element)
+    if elm_car:
+        db.session.delete(elm_car)
+
     db.session.commit()
     flash('Elemento eliminado', 'success')
     return redirect(url_for('inventario'))
@@ -413,7 +441,7 @@ def filtrar_elementos():
 
     # Iniciar una consulta base para los elementos
     query = db.session.query(Element)
-
+    filtro=[0]
     # Aplicar filtro por nombre si está presente
     if name_filter:
         query = query.filter(Element.name.ilike(f"%{name_filter}%"))
@@ -421,17 +449,23 @@ def filtrar_elementos():
     # Aplicar filtro por categorías si hay categorías seleccionadas
     if categories_filter:
         category_ids = [int(c) for c in categories_filter.split(',')]
+
+        filtro = Category.query.filter(Category.IDcategory.in_(category_ids))
+        filtro = filtro.all()
+
         query = query.join(ElementCategory).filter(ElementCategory.IDcategory.in_(category_ids))
 
     # Ejecutar la consulta y obtener los elementos que coinciden con los filtros
     elementos_filtrados = query.all()
-
+    
     # Obtener todas las categorías para mostrarlas en el filtro
     categorias = Category.query.all()
 
     # Renderizar la plantilla con los elementos filtrados
-
-    return render_template('inventario/index.html', categories=categorias, elements=elementos_filtrados)
+    if filtro:
+        return render_template('inventario.html', categories=categorias, elements=elementos_filtrados,filtro=filtro)
+    
+    return render_template('inventario.html', categories=categorias, elements=elementos_filtrados)
 
 ########### /////////// #########
 
@@ -448,8 +482,131 @@ def localizacion():
 
 ########### /////////// #########
 
+@app.route('/add_location', methods=['POST'])
+def add_location():
+    # Obtener datos del FormData
+    name = request.form.get('name')
+    planta = request.form.get('floor')
+    description = request.form.get('description')
+    image = request.files.get('image')
+
+    # Verificar si los campos obligatorios están presentes
+    if not name:
+        return jsonify({'message': 'Faltan campos obligatorios.'}), 400
+
+    # Verificar si el archivo es permitido
+    if not allowed_file(image.filename):
+        return jsonify({'message': 'Tipo de archivo no permitido.'}), 400
+
+    # Guardar la imagen de manera segura
+    filename = secure_filename(image.filename)
+    image_path = os.path.join(app.config['UPLOAD_FOLDER'], filename)
+    image.save(image_path)
+    image_path='static/uploads/'+filename
+    # Crear el nuevo elemento
+    nueva_localizacion = Location(name=name, planta=planta, descrip=description, img=filename)
+
+    # Añadir a la base de datos
+    db.session.add(nueva_localizacion)
+    db.session.commit()
 
 
+    return jsonify({'message': 'Localizacion cargada.'}), 200
+
+@app.route('/borrar_location/<int:id>', methods=['POST'])
+def borrar_location(id):
+    location = Location.query.get_or_404(id)
+
+    if location.img:
+                    try:
+                        os.remove(os.path.join(app.config['UPLOAD_FOLDER'], location.img))
+                    except FileNotFoundError:
+                        pass 
+    
+    db.session.delete(location)
+    db.session.commit()
+    flash('Localizacion eliminada', 'success')
+    return redirect(url_for('localizacion'))
+############################ ////////////// ###################################################
+
+
+
+############################ Locura maximaaaa ###################################################
+
+@app.route('/elements', methods=['GET'])
+def get_elements():
+    categories = request.args.getlist('categories[]')  # Filtros de categorías seleccionados
+    query = db.session.query(Element)  # Empezamos la consulta con los elementos
+    
+    if categories:  # Si hay filtros de categoría seleccionados
+        query = query.join(ElementCategory).filter(ElementCategory.IDcategory.in_(categories))
+
+    # Ordenamos por disponibilidad y por ID de elemento
+    elements = query.order_by(Element.disp.desc(), Element.IDelement).all()
+
+    # Construimos la lista de elementos con la información que necesitamos
+    elements_data = [{'IDelement': e.IDelement, 'name': e.name, 'disp': e.disp, 'candisp': e.candisp, 'img': e.img} for e in elements]
+    
+    return jsonify(elements_data)
+
+@app.route('/categories', methods=['GET'])
+def get_categories():
+    categories = Category.query.all()
+    categories_data = [{'IDcategory': c.IDcategory, 'name': c.name} for c in categories]
+    return jsonify(categories_data)
+
+
+@app.route('/locations', methods=['GET'])
+def get_locations():
+    locations = Location.query.all()
+    locations_data = [{'id': loc.id, 'name': loc.name} for loc in locations]
+    return jsonify(locations_data)
+
+############### carrito ###############
+#
+
+
+@app.route('/checkout', methods=['GET', 'POST'])
+def checkout():
+    if request.method == 'GET':
+        # Renderizar la página de checkout con la lista de elementos seleccionados
+        return render_template('PEDIDOS/carrito.html')
+    
+    if request.method == 'POST':
+        data = request.json  # Datos enviados desde el frontend (elementos seleccionados, cantidades, etc.)
+        IDuser = current_user.IDuser # Obtener IDuser desde sesión o autenticación (esto es un ejemplo)
+
+        # Convertir la fecha enviada como string a un objeto datetime.date
+        order_date = datetime.strptime(data['date'], '%Y-%m-%d').date()
+
+        # Convertir la hora enviada como string a un objeto datetime.time
+        order_time = datetime.strptime(data['time'], '%H:%M').time()  # Asegúrate de que 'time' sea parte de los datos enviados
+
+        # Crear la orden con la localización seleccionada
+        new_order = Order(
+            IDuser=IDuser,
+            date=order_date,  # Pasamos la fecha como un objeto date de Python
+            time=order_time,
+            location=data['location'],  # Localización seleccionada por el usuario
+            repeat = data['repeat'],
+            state='pending'
+        )
+        db.session.add(new_order)
+        db.session.flush()  # Para obtener el ID de la orden recién creada
+
+        # Insertar los elementos del pedido en la tabla OrderElement
+        for item in data['items']:
+            order_element = OrderElement(IDorder=new_order.IDorder, IDelement=item['IDelement'], cant=item['quantity'])
+            db.session.add(order_element)
+
+        # Si se seleccionó "Repetir pedido", guardar los días de repetición
+        if data.get('repeat'):
+            for day in data['repeatDays']:
+                order_rep = OrderRep(IDorder=new_order.IDorder, repeat_day=day)
+                db.session.add(order_rep)
+
+        db.session.commit()
+        return jsonify({'success': True})
 
 ############################ ////////////// ###################################################
